requirements.txt
PyYAML>=6.0
pycdlib>=1.14.0
